--- a/lib/vm/src/trap/mod.rs
+++ b/lib/vm/src/trap/mod.rs
@@ -9,6 +9,6 @@ mod traphandlers;
 pub use trapcode::TrapCode;
 pub use traphandlers::{
     catch_traps, catch_traps_with_result, raise_lib_trap, raise_user_trap, wasmer_call_trampoline,
-    TlsRestore, Trap, TrapHandler, TrapHandlerFn,
+    wasmer_call_trampoline_unchecked, TlsRestore, Trap, TrapHandler, TrapHandlerFn,
 };
 pub use traphandlers::{init_traps, resume_panic};