<!-- 
Prior to submitting a PR, review the CONTRIBUTING.md document for recommendations on how to test:
https://github.com/wasmerio/wasmer/blob/master/CONTRIBUTING.md#pull-requests

-->

# Description

This PR resolves #1922 by adding a new method to `Store`: `catch_traps` and two new methods to `Function` and `NativeFunction`: `call_unchecked` so unsafe unchecked calls can be done.

```rust
unsafe {
  store.catch_traps(|| {
    assert_eq!(sum.call_unchecked(&[Value::I32(1), Value::I32(2)]).unwrap().to_vec(), vec![Value::I32(3)]);
  });
}
```

## Benchmarking

This are the timings in of `make bench` in my laptop (macOS M1 2020):
* Static function: 138ns -> 3ns
* Dynamic function: 240ns -> 100ns

```
basic static func llvm
                        time:   [136.39 ns 137.16 ns 138.08 ns]

basic static func llvm [unchecked]
                        time:   [3.0288 ns 3.0381 ns 3.0485 ns]

basic static func with many args llvm
                        time:   [149.08 ns 149.28 ns 149.50 ns]

basic static func with many args llvm [unchecked]
                        time:   [15.919 ns 16.085 ns 16.290 ns]

basic dynfunc llvm
                        time:   [242.52 ns 244.49 ns 247.13 ns]

basic dynfunc llvm [unchecked]
                        time:   [101.92 ns 103.36 ns 104.93 ns]
```