@syrusakbary 2021-05-17
> What if we improve that, and remove this “unsafe API”, and replace it by a new “batch API”. It's probably a lot more thinking, but that's really what we are trying to achieve here!
> 
> We could simply provide an API like:
> 
```rust
store.batch_functions(|| {
    function1.call(…);
    do something else
    function2.call(…);
})
```

`catch_traps` and `call_unchecked` are highly unsafe.
If the `do something else` code has to run any destructor, and `function2.call` traps... the destructor will not be run. (this is documented in the `#Safety` section in the docs for `call_unchecked`.