@Hywan 2021-05-18
Then we can mark `Store::batch_functions` as `unsafe`. But the idea is to create an API that reflects the intent, what it is designed for, for the users. Here, the intent is to allow multiple calls.

If “batch” is too specific, we could do something have a more “generic” API like “fast calls”, which will continue to be unsafe of course, but it's better than “unchecked” from my point of view.

Thoughts?