@heyjdp 2022-05-24
Closing PR, this will be solved by Wasmer v3.0